refactor attester slashing schema building
